requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
drf_yasg
requests
django-filter
pillow
django-map-widgets
